requirements.txt
certifi==2019.11.28
chardet==3.0.4
gitdb2==2.0.6
GitPython==3.0.7
idna==2.8
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
smmap2==2.0.5
urllib3==1.25.8
